Written slab_cache_init()
