+++ kernel/src/mm/slab.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014 DarkSide Project
+ * Copyright (C) 2014-2015 DarkSide Project
  * Authored by George Klees <gksharkboy@gmail.com>
  * slab.c - Slab cache management
  *
@@ -24,33 +24,27 @@
 #include <mm/addrspace.h>
 #include <mm/slab.h>
 
-/* Create a slab cache */
-slab_cache_t *slab_cache_create(size_t object_size, int flags)
+/* Initialize a slab cache without any allocations */
+void slab_cache_init(slab_cache_t *slab_cache, void *slab, size_t object_size, int flags)
 {
-	slab_cache_t *slab_cache = addrspace_alloc(ADDRSPACE_SYSTEM, SLAB_SIZE, SLAB_SIZE, flags | PAGE_GLOBAL | PAGE_PRIVATE);
-	slab_cache_init(slab_cache, object_size, flags);
-	return slab_cache;
-}
-
-/* Initialize a slab cache */
-void slab_cache_init(slab_cache_t *slab_cache, size_t object_size, int flags)
-{
-	/* Fill in basic information */
+	/* Fill in the slab cache information */
 	slab_cache->object_size = object_size;
 	slab_cache->flags = flags;
+	slab_cache->empty = (slab_header_t*) slab;
+	slab_cache->partial = NULL;
+	slab_cache->full = NULL;
 	spinlock_recursive_init(&slab_cache->lock);
-	slab_cache->next = NULL;
 
 	/* Calculate the amount of available space in the slab */
-	size_t available_space = SLAB_SIZE - sizeof(slab_cache_t);
+	size_t available_space = SLAB_SIZE - sizeof(slab_header_t);
 
 	/* Find the largest number of objects we can pack into the slab */
-	size_t num_total_objs = 1;
+	size_t objs_per_slab = 0;
 	while (true)
 	{
 		/* Calculate the space needed for objects and bitmaps */
-		size_t object_space = num_total_objs * object_size;
-		size_t bitmap_space = ceil(num_total_objs, 8);
+		size_t object_space = (num_total_objs + 1) * object_size;
+		size_t bitmap_space = ceil(num_total_objs + 1, 8);
 
 		/* Adjust the bitmap space to a multiple of 4, if needed */
 		if (bitmap_space & 3)
@@ -61,117 +55,34 @@ void slab_cache_init(slab_cache_t *slab_cache, size_t object_size, int flags)
 		/* If the needed space exceeds the available space, end the loop */
 		if (object_space + bitmap_space > available_space)
 		{
-			/* Fill in the remaining data */
-			slab_cache->object_data_start = sizeof(slab_cache_t) + bitmap_space;
-			slab_cache->num_total_objs = slab_cache->num_free_objs = num_total_objs;
+			/* We can't put any objects in the slab */
+			if (objs_per_slab == 0) return;
+
+			/* Fill in the slab header size and number of objects per slab */
+			slab_cache->slab_header_size = sizeof(slab_header_t) + bitmap_space;
+			slab_cache->objs_per_slab = objs_per_slab;
+
+			/* Set up the slab's data */
+			slab_header_t *slab_header = (slab_header_t*) slab;
+			slab_header->num_free_objs = objs_per_slab;
+			slab_header->next = NULL;
+			spinlock_recursive_init(&slab_header->lock);
 
 			/* Mark the entire bitmap as used */
-			memset(slab_cache->free_bitmap, 0xFF, bitmap_space);
+			memset(slab_header->free_bitmap, 0xFF, bitmap_space);
 
 			/* Free objects that are in both the slab and bitmap */
-			for (size_t i = 0; i < num_total_objs; i++)
+			for (size_t i = 0; i < objs_per_slab; i++)
 			{
 				size_t byte_start = i / 8;
 				uint8_t bit_start = i % 8;
-				slab_cache->free_bitmap[byte_start] &= ~(1 << bit_start);
+				slab_header->free_bitmap[byte_start] &= ~(1 << bit_start);
 			}
 
 			return;
 		}
 
-		num_total_objs++;
-	}
-}
-
-/* Allocate an object from a slab cache */
-void *slab_cache_alloc(slab_cache_t *slab_cache)
-{
-allocation: ;
-	/* Iterate through each slab cache */
-	slab_cache_t *prev_slab_cache = NULL;
-	while (slab_cache)
-	{
-		/* Acquire the lock on the slab cache */
-		spinlock_recursive_acquire(&slab_cache->lock);
-
-		/* If the number of free objects is 0, go to the next cache */
-		if (slab_cache->num_free_objs == 0)
-		{
-			spinlock_recursive_release(&slab_cache->lock);
-			prev_slab_cache = slab_cache;
-			slab_cache = slab_cache->next;
-			continue;
-		}
-
-		/* Search the bitmap for an available free object */
-		for (uint32_t i = 0; i < ceil(slab_cache->num_total_objs, 8); i++)
-		{
-			uint8_t byte = slab_cache->free_bitmap[i];
-
-			for (uint8_t j = 0; j < 8; j++)
-			{
-				if (byte == 0xFF)
-				{
-					break;
-				}
-
-				if (byte & 1)
-				{
-					byte >>= 1;
-					continue;
-				}
-
-				slab_cache->free_bitmap[i] |= (1 << j);
-				slab_cache->num_free_objs--;
-
-				size_t object_num = (i * 8) + j;
-				void *object = ((void*) slab_cache) + slab_cache->object_data_start + (object_num * slab_cache->object_size);
-				spinlock_recursive_release(&slab_cache->lock);
-				return object;
-			}
-		}
-	}
-
-	/* Acquire the lock on the previous (last valid) slab cache */
-	spinlock_recursive_acquire(&prev_slab_cache->lock);
-
-	/* Allocate a new slab and initialize it */
-	slab_cache = slab_cache_create(prev_slab_cache->object_size, prev_slab_cache->flags);
-	prev_slab_cache->next = slab_cache;
-
-	/* Perform the allocation again */
-	spinlock_recursive_release(&prev_slab_cache->lock);
-	goto allocation;
-}
-
-/* Free an object from a slab cache */
-void slab_cache_free(slab_cache_t *slab_cache, void *ptr)
-{
-	/* Iterate through each slab cache */
-	while (slab_cache)
-	{
-		/* If we're not in range, go to the next slab cache */
-		if (ptr < (void*) slab_cache || ptr > (((void*) slab_cache) + SLAB_SIZE))
-		{
-			slab_cache = slab_cache->next;
-			continue;
-		}
-
-		/* Acquire the lock on the slab cache */
-		spinlock_recursive_acquire(&slab_cache->lock);
-
-		/* Clear the object's location in the bitmap and increment the number of free objects */
-		size_t object_num = (size_t) (ptr - ((void*) slab_cache) - slab_cache->object_data_start) / slab_cache->object_size;
-		if (object_num < slab_cache->num_total_objs)
-		{
-			size_t byte_start = object_num / 8;
-			uint8_t bit_start = object_num % 8;
-			slab_cache->free_bitmap[byte_start] &= ~(1 << bit_start);
-		}
-		slab_cache->num_free_objs++;
-
-		/* Release the lock and return */
-		spinlock_recursive_release(&slab_cache->lock);
-		return;
+		/* We can fit one more object than our original guess */
+		objs_per_slab++;
 	}
 }
